requirements.txt
numpy>=1.20
Pillow>=9.0
matplotlib>=3.5
